+++ functionGenerator.h
@@ -2,14 +2,14 @@
 //
 //    FILE: functionGenerator.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.7
+// VERSION: 0.3.0
 // PURPOSE: wave form generating functions (use with care)
 //     URL: https://github.com/RobTillaart/FunctionGenerator
 
 
 #include "Arduino.h"
 
-#define FUNCTIONGENERATOR_LIB_VERSION           (F("0.2.7"))
+#define FUNCTIONGENERATOR_LIB_VERSION           (F("0.3.0"))
 
 
 class funcgen
@@ -71,7 +71,7 @@ public:
   float trapezium1(float t);
   float trapezium2(float t);
   float heartBeat(float t);  //  72 BPM = 72/60 = 1 setFrequency(1.2)
-  float freeWave(float t, int16_t * arr);
+  float freeWave(float t, int16_t * arr, int16_t N);  //  arr must be N+1 long
 
 
 private: